+++ src/components/AsteroidWithRef.js
@@ -1,0 +1,62 @@
+// src/components/AsteroidWithRef.js
+import React, { useRef, useEffect, useMemo } from 'react';
+import { useFrame } from '@react-three/fiber';
+import * as THREE from 'three';
+
+function AsteroidWithRef({
+  asteroidData,
+  speedMultiplier,
+  setSelectedObject,
+  setHoveredObject,
+  setAsteroidRefs,
+}) {
+  const asteroidRef = useRef();
+  const orbitRadius =
+    parseFloat(asteroidData.close_approach_data[0].miss_distance.kilometers) /
+    100000;
+  const orbitSpeed = 0.05 / (orbitRadius / 10);
+  const initialAngle = useMemo(() => Math.random() * 2 * Math.PI, []);
+
+  const handleAsteroidClick = () => {
+    setSelectedObject({ ref: asteroidRef, data: asteroidData });
+  };
+
+  const handlePointerOver = () => {
+    setHoveredObject(asteroidData.name);
+    document.body.style.cursor = 'pointer';
+  };
+
+  const handlePointerOut = () => {
+    setHoveredObject(null);
+    document.body.style.cursor = 'auto';
+  };
+
+  useEffect(() => {
+    setAsteroidRefs((prevRefs) => ({
+      ...prevRefs,
+      [asteroidData.name]: asteroidRef,
+    }));
+  }, [asteroidData.name, setAsteroidRefs]);
+
+  useFrame(({ clock }) => {
+    const time = clock.getElapsedTime() * speedMultiplier;
+    const orbitPosition = initialAngle + time * orbitSpeed;
+    const x = orbitRadius * Math.cos(orbitPosition);
+    const z = orbitRadius * Math.sin(orbitPosition);
+    asteroidRef.current.position.set(x, 0, z);
+  });
+
+  return (
+    <mesh
+      ref={asteroidRef}
+      onClick={handleAsteroidClick}
+      onPointerOver={handlePointerOver}
+      onPointerOut={handlePointerOut}
+    >
+      <sphereGeometry args={[0.1, 8, 8]} />
+      <meshStandardMaterial color={'white'} />
+    </mesh>
+  );
+}
+
+export default AsteroidWithRef;
