+++ src/components/Orbit.js
@@ -1,0 +1,33 @@
+// src/components/Orbit.js
+import React, { useMemo } from 'react';
+import * as THREE from 'three';
+
+function Orbit({ radius }) {
+  const points = useMemo(() => {
+    const temp = [];
+    for (let i = 0; i <= 64; i++) {
+      const angle = (i / 64) * 2 * Math.PI;
+      temp.push(
+        new THREE.Vector3(
+          Math.cos(angle) * radius,
+          0,
+          Math.sin(angle) * radius
+        )
+      );
+    }
+    return temp;
+  }, [radius]);
+
+  const orbitGeometry = useMemo(
+    () => new THREE.BufferGeometry().setFromPoints(points),
+    [points]
+  );
+
+  return (
+    <line geometry={orbitGeometry}>
+      <lineBasicMaterial color="white" transparent={true} opacity={0.3} />
+    </line>
+  );
+}
+
+export default Orbit;
